poc/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
geoalchemy2==0.14.6
python-dotenv==1.0.1
razorpay==1.4.1
twilio==9.1.0
firebase-admin==6.5.0
httpx==0.27.0
langchain==0.2.1
langchain-anthropic==0.1.13
scikit-learn==1.4.2
pandas==2.2.2
redis==5.0.4
